requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
sqlalchemy>=2.0.19
mysql-connector-python>=8.0.33
pydantic>=2.1.1
python-dotenv>=1.0.0
streamlit>=1.25.0
pandas>=2.0.3
plotly>=5.15.0
python-decouple>=3.8
email-validator>=2.0.0
